+++ src/services/@types/favorites.ts
@@ -13,10 +13,19 @@
 //	Exports ____________________________________________________________________
 
 export type Favorite = {
-	fileA:string,
-	fileB:string,
-	label:string
+	fileA:string;
+	fileB:string;
+	label:string;
+	groupId?:number;
 };
+
+export type FavoriteGroup = {
+	label:string;
+	id:number;
+	collapsed:boolean;
+};
+
+export type InitialState = 'Collapsed'|'Expanded'|'Remember';
 
 //	Functions __________________________________________________________________