+++ src/services/panel/DiffPanel.ts
@@ -57,6 +57,9 @@ export class DiffPanel {
 	
 	private disposables:vscode.Disposable[] = [];
 	
+	private _onDidInit:vscode.EventEmitter<undefined> = new vscode.EventEmitter<undefined>();
+	public readonly onDidInit:vscode.Event<undefined> = this._onDidInit.event;
+	
 	private constructor (panel:vscode.WebviewPanel, context:vscode.ExtensionContext, uris:null|Uri[]|vscode.Uri[] = null, compare?:boolean) {
 		
 		this.panel = panel;
@@ -203,7 +206,7 @@ export class DiffPanel {
 		
 		this.copy.onDidCopyFile(({ from, to }) => {
 			
-			this.output.log(`Copied ${from.type} "${from.name}" from "${from.folder}" to "${to.folder}".`);
+			this.output.log(`Copied ${from.type} "${from.name}" from "${from.root}" to "${to.root}".`);
 			
 		}, null, this.disposables);
 		
@@ -258,9 +261,9 @@ export class DiffPanel {
 		
 		this.msg.on('open:dialog', async () => {
 			
-			const folder = await DiffDialog.open();
+			const fsPath = await DiffDialog.open();
 			
-			this.msg.send('open:dialog', { folder });
+			this.msg.send('open:dialog', { fsPath });
 			
 		});
 		
@@ -295,6 +298,7 @@ export class DiffPanel {
 			});
 			
 			this.msg.removeMessageListener('init:view');
+			this._onDidInit.fire(undefined);
 			
 		});
 		
@@ -334,7 +338,7 @@ export class DiffPanel {
 	
 	private setTitle (pathA?:string, pathB?:string) {
 		
-		let title = 'Diff';
+		let title = 'Diff Folders';
 		
 		if (pathA && pathB) {
 			const [label, desc] = formatNameAndDesc(pathA, pathB);
@@ -380,7 +384,17 @@ export class DiffPanel {
 				<meta charset="UTF-8">
 				<meta http-equiv="Content-Security-Policy" content="${csp}">
 				<meta name="viewport" content="width=device-width, initial-scale=1.0">
-				<title>L13 Diff</title>
+				<title>Diff Folders</title>
+				<script nonce="${nonceToken}">
+					const timeoutId = setTimeout(() => {
+						
+						acquireVsCodeApi().postMessage({
+							command: 'error:init',
+							data: { error: 'Failed to load resources!' },
+						});
+						
+					}, 1000);
+				</script>
 				<link rel="stylesheet" nonce="${nonceToken}" href="${webview.asWebviewUri(styleUri)}">
 				<script nonce="${nonceToken}" src="${webview.asWebviewUri(scriptUri)}"></script>
 			</head>
@@ -389,9 +403,9 @@ export class DiffPanel {
 		
 	}
 	
-	public static create (context:vscode.ExtensionContext, uris:null|Uri[]|vscode.Uri[] = null, compare?:boolean) {
+	public static async create (context:vscode.ExtensionContext, uris:null|Uri[]|vscode.Uri[] = null, compare?:boolean) {
 		
-		const panel = vscode.window.createWebviewPanel(DiffPanel.viewType, 'Diff', {
+		const panel = vscode.window.createWebviewPanel(DiffPanel.viewType, 'Diff Folders', {
 			viewColumn: vscode.ViewColumn.Active,
 		}, {
 			enableScripts: true,
@@ -410,6 +424,22 @@ export class DiffPanel {
 		
 		DiffPanel.currentPanel = diffPanel;
 		DiffPanel.currentPanels.push(diffPanel);
+		
+		return new Promise((resolve, reject) => {
+			
+			diffPanel.onDidInit(() => resolve(), null, diffPanel.disposables);
+			
+			diffPanel.msg.on('error:init', async ({ error }) => {
+				
+				const tryAgain = await vscode.window.showErrorMessage(error, 'Try Again');
+				
+				diffPanel.dispose();
+				
+				resolve(tryAgain ? DiffPanel.create(context, uris, compare) : undefined);
+				
+			});
+			
+		});
 		
 	}